refactor: extract selection state from AgentWorkspace
